LayoutTests/inspector-protocol/debugger/setPauseOnExceptions-all-expected.txt

CONSOLE MESSAGE: line 1: caught inline: {"stack":"","line":1}
CONSOLE MESSAGE: line 38: catchNested caught exception: {"stack":"exceptionBasic@exception.js"}
CONSOLE MESSAGE: line 38: catchNested caught exception: {"code":8,"message":"NotFoundError: DOM Exception 8","name":"NotFoundError","stack":"appendChild@[native code]\nexceptionDOM@exception.js"}
CONSOLE MESSAGE: line 38: catchNested caught exception: "exception in host function"
CONSOLE MESSAGE: line 38: catchNested caught exception: "exception string"
CONSOLE MESSAGE: line 38: catchNested caught exception: {"line":18,"sourceURL":"exception.js:33:26\n"}
CONSOLE MESSAGE: line 1: TypeError: 'undefined' is not an object (evaluating '({}).a.b')
CONSOLE MESSAGE: line 3: TypeError: 'undefined' is not an object (evaluating '({}).a.b')
CONSOLE MESSAGE: line 8: NotFoundError: DOM Exception 8: An attempt was made to reference a Node in a context where it does not exist.
CONSOLE MESSAGE: exception in host function
CONSOLE MESSAGE: exception string
CONSOLE MESSAGE: line 18: [object Object]
CONSOLE MESSAGE: line 18: Error: error message
Debugger.setPauseOnExceptions all

Breakpoints Enabled
Debugger.setPauseOnExceptions - all
PASS - did not pause on non-exception statements
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
PASS - paused for each caught exception
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
Paused!
Resumed
PASS - paused for each uncaught exception
PASS - did not pause on non-exception statements
